# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmAid API
  description: Disaster relief coordination platform connecting surplus resources with needs
  version: 1.0.0
  contact:
    name: SwarmAid Team
    email: api@swarmaid.org
servers:
  - url: https://api.swarmaid.org
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
    Coordinates:
      type: object
      properties:
        lat:
          type: number
        lon:
          type: number
      required: [lat, lon]
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
      required: [street, city, state, postalCode, country]
    TimeWindow:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
      required: [start, end]
    SurplusListing:
      type: object
      properties:
        id:
          type: string
        supplierId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [perishable_food, non_perishable_food, medical_supplies, hygiene_products, clothing, shelter_materials, equipment, other]
        quantity:
          type: number
        quantityUnit:
          type: string
        estimatedValue:
          type: number
        images:
          type: array
          items:
            type: string
        pickupAddress:
          $ref: '#/components/schemas/Address'
        pickupCoordinates:
          $ref: '#/components/schemas/Coordinates'
        pickupWindow:
          $ref: '#/components/schemas/TimeWindow'
        expirationDate:
          type: string
          format: date-time
        requiresRefrigeration:
          type: boolean
        handlingRequirements:
          type: array
          items:
            type: string
        qualityNotes:
          type: string
        status:
          type: string
          enum: [posted, matched, scheduled, picked_up, delivered, canceled, expired, failed]
        enrichment:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /supply:
    post:
      summary: Create surplus listing
      tags: [Supply]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                quantity:
                  type: number
                quantityUnit:
                  type: string
                estimatedValue:
                  type: number
                pickupAddress:
                  $ref: '#/components/schemas/Address'
                pickupWindow:
                  $ref: '#/components/schemas/TimeWindow'
                expirationDate:
                  type: string
                  format: date-time
                requiresRefrigeration:
                  type: boolean
                handlingRequirements:
                  type: array
                  items:
                    type: string
                qualityNotes:
                  type: string
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurplusListing'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List surplus listings
      tags: [Supply]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: '#/components/schemas/SurplusListing'
                  nextToken:
                    type: string
...